+++ handler.go
@@ -5,107 +5,76 @@
 package main
 
 import (
+	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
-	"os"
 	"path/filepath"
 	"time"
 )
 
-func notFound(w http.ResponseWriter, err error) {
-	errString := ""
-	if err != nil {
-		errString = err.Error()
-	}
-
-	http.Error(w, errString, http.StatusNotFound)
+func notFoundError(w http.ResponseWriter, err error) {
+	http.Error(w, err.Error(), http.StatusNotFound)
 }
 
-func badRequest(w http.ResponseWriter, err error) {
-	errString := ""
-	if err != nil {
-		errString = err.Error()
-	}
-
-	http.Error(w, errString, http.StatusBadRequest)
+func badRequest(w http.ResponseWriter, errorMessage string) {
+	http.Error(w, errorMessage, http.StatusBadRequest)
 }
 
-func catchAll(publicPath string, cacheTime time.Duration) func(w http.ResponseWriter, r *http.Request) {
+func catchAll(public http.FileSystem, cacheTime time.Duration) func(w http.ResponseWriter, r *http.Request) {
+	fileServer := http.FileServer(public)
 	return func(w http.ResponseWriter, r *http.Request) {
-
-		path := r.RequestURI
-
-		file, err := os.Stat(publicPath + path)
-		if os.IsNotExist(err) {
-			notFound(w, fmt.Errorf("404 Page Not Found Error"))
-			return
-		}
-
-		// Handles if it's a directory or last char is a / (also a directory)
-		// It then opens index.json by default (if it exists)
-		if len(path) == 0 {
-			path = "/index.json"
-		} else if path[len(path)-1:] == "/" {
-			path = path + "index.json"
-		} else if file.IsDir() {
-			path = path + "/index.json"
-		}
-
-		file, err = os.Stat(publicPath + path)
-		if os.IsNotExist(err) {
-			notFound(w, fmt.Errorf("404 Page Not Found Error"))
-			return
-		}
-
-		data, err := ioutil.ReadFile(publicPath + path)
+		path, err := determineResourcePath(r, public)
 		if err != nil {
-			notFound(w, fmt.Errorf("404 Page Not Found Error"))
+			notFoundError(w, err)
 			return
 		}
 
 		cacheHeaders(w, cacheTime)
-		sendHeader(w, r)
-		w.Write(data)
+
+		r.URL.Path = path
+		fileServer.ServeHTTP(w, r)
 	}
 }
 
-func jsonHeader(w http.ResponseWriter) {
-	w.Header().Set("Content-Type", "application/json")
-}
+func determineResourcePath(r *http.Request, public http.FileSystem) (string, error) {
+	path := r.URL.Path
 
-func sendHeader(w http.ResponseWriter, r *http.Request) {
-	extension := filepath.Ext(r.RequestURI)
+	// Handles if it's a directory or last char is a / (also a directory)
+	// It then opens index.json by default (if it exists)
+	if len(path) == 0 || path == "/" {
+		path = "index.json"
+	} else if path[len(path)-1:] == "/" {
+		path = filepath.Join(path, "index.json")
+	} else {
+		f, err := public.Open(path)
+		if err != nil { // catch all errors, including "forbidden access"
+			return "", errors.New("404 Page Not Found Error")
+		}
+		defer f.Close()
 
-	switch extension {
-	// No extension is always json
-	case "":
-		w.Header().Set("Content-Type", "application/json")
-	case ".asciidoc":
-		w.Header().Set("Content-Type", "text/asciidoc; charset=UTF-8")
-	case ".gz":
-		w.Header().Set("Content-Type", "application/gzip")
-	case ".ico":
-		w.Header().Set("Content-Type", "image/x-icon")
-	case ".jpg":
-	case ".jpeg":
-		w.Header().Set("Content-Type", "image/jpeg")
-	case ".md":
-		w.Header().Set("Content-Type", "text/markdown; charset=UTF-8")
-	case ".png":
-		w.Header().Set("Content-Type", "image/png")
-	case ".svg":
-		w.Header().Set("Content-Type", "image/svg+xml")
-	case ".yml":
-		w.Header().Set("Content-Type", "text/yaml; charset=UTF-8")
-	default:
-		// Using json as the default header
-		w.Header().Set("Content-Type", "application/json")
+		stat, err := f.Stat()
+		if err != nil {
+			return "", errors.New("404 Page Not Found Error")
+		}
+
+		if stat.IsDir() {
+			path = path + "/index.json"
+			dirIndexFile, err := public.Open(path)
+			if err != nil { // catch all errors, including "forbidden access"
+				return "", errors.New("404 Page Not Found Error")
+			}
+			defer dirIndexFile.Close()
+		}
 	}
+	return path, nil
 }
 
 func cacheHeaders(w http.ResponseWriter, cacheTime time.Duration) {
 	maxAge := fmt.Sprintf("max-age=%.0f", cacheTime.Seconds())
 	w.Header().Add("Cache-Control", maxAge)
 	w.Header().Add("Cache-Control", "public")
+}
+
+func jsonHeader(w http.ResponseWriter) {
+	w.Header().Set("Content-Type", "application/json")
 }